Adds a tag to berksfile
It is dangerous to use berkshelf against a git repo since berks will always fetch the latest release. It also misleads the user into thinking that they have a specific cookbook version, when really the cookbook version is whatever HEAD was set to on the last fetch. It is safer to recommend users pin to a specific tag.